+++ packages/avatars-human-sprites/src/index.ts
@@ -6,10 +6,12 @@ type Options = {
   mood?: Array<'happy' | 'sad' | 'surprised'>;
 };
 
-export default function(random: Random, options: Options = {}) {
+export function create(random: Random, options: Options = {}) {
   if (random.bool(50)) {
     return Male(random, options);
   } else {
     return Female(random, options);
   }
 }
+
+export default create;