+++ packages/avatars-jdenticon-sprites/src/index.ts
@@ -10,21 +10,20 @@ type Options = {
   grayscaleSaturation?: number;
 };
 
-export default function(random: Random, options: Options = {}) {
+export function create(random: Random, options: Options = {}) {
   jdenticon.config = {
     hues: options.hues,
     lightness: {
       color: options.colorLightness,
-      grayscale: options.grayscaleLightness
+      grayscale: options.grayscaleLightness,
     },
     saturation: {
       color: options.colorSaturation,
-      grayscale: options.grayscaleSaturation
-    }
+      grayscale: options.grayscaleSaturation,
+    },
   };
 
-  return jdenticon
-    .toSvg(random.seed, 50, 0)
-    .replace('width="50"', '')
-    .replace('height="50"', '');
+  return jdenticon.toSvg(random.seed, 50, 0).replace('width="50"', '').replace('height="50"', '');
 }
+
+export default create;